+++ checkpoint/orbax/checkpoint/composite_checkpoint_handler.py
@@ -56,6 +56,7 @@ import uuid
 
 from absl import logging
 from etils import epath
+import jax
 import nest_asyncio
 from orbax.checkpoint import async_checkpoint_handler
 from orbax.checkpoint import checkpoint_args
@@ -376,7 +377,6 @@ class CompositeCheckpointHandler(AsyncCheckpointHandler):
       self, directory: epath.Path, args: 'CompositeArgs'
   ) -> Optional[List[Future]]:
     """Saves multiple items to individual subdirectories."""
-    futures = []
     # Sort keys to maintain consistent ordering across processes, otherwise
     # we may hit timeouts if processes wait at different barriers in per-item
     # handlers.
@@ -399,19 +399,19 @@ class CompositeCheckpointHandler(AsyncCheckpointHandler):
     for path in self._current_temporary_paths.values():
       path.create()
 
+    save_ops = []
     for item_name, item_directory in self._current_temporary_paths.items():
       arg = args[item_name]
       _maybe_raise_reserved_item_error(item_name)
       handler = self._get_or_set_handler(item_name, arg)
       if isinstance(handler, AsyncCheckpointHandler):
-        commit_futures = await handler.async_save(
-            item_directory.get(), args=arg
-        )
-        if commit_futures is not None:
-          futures.extend(commit_futures)
+        save_ops.append(handler.async_save(item_directory.get(), args=arg))
       else:
+        # Blocking save.
         handler.save(item_directory.get(), args=arg)
-    return futures
+
+    commit_futures = jax.tree.flatten(await asyncio.gather(*save_ops))[0]
+    return commit_futures or []
 
   def save(self, *args, **kwargs):
     """Saves synchronously."""